requirements.txt
streamlit
google-generativeai
requests
yfinance
scipy
matplotlib
numpy
pandas